vendor_service/requirements.txt
black
fastapi
flake8
python-dotenv
uvicorn
requests
